+++ tests/event_lists/globals/test_all_russia.py
@@ -430,7 +430,6 @@ data1 = {
     "Women's national sports teams of Russia": "منتخبات رياضية وطنية نسائية في روسيا",
     "Works about the French invasion of Russia": "أعمال عن الغزو الفرنسي لروسيا",
     "Films about the French invasion of Russia": "أفلام عن الغزو الفرنسي لروسيا",
-
     "Agriculture ministers of Russia": "وزراء زراعة روسيا",
     "Culture ministers of Russia": "وزراء ثقافة روسيا",
     "Economy ministers of Russia": "وزراء اقتصاد روسيا",