+++ tests/test_skip_data.py
@@ -44,7 +44,6 @@ executed_data_1 = {
 }
 
 fast_data_1 = {
-
     "Jews and Judaism in Kingdom of Jerusalem": "اليهود واليهودية في مملكة بيت المقدس",
     "Works about foreign relations of the United States": "أعمال عن علاقات خارجية الولايات المتحدة",
     "Books about foreign relations of the United States": "كتب عن علاقات خارجية الولايات المتحدة",
@@ -75,6 +74,7 @@ fast_data_1 = {
 to_test = [
     ("test_sports_events_2", fast_data_1),
 ]
+
 
 @pytest.mark.parametrize("category, expected", fast_data_1.items(), ids=fast_data_1.keys())
 @pytest.mark.skip2